fix: 빌드 에러 수정

+++ backend/backend-user/src/main/java/com/example/user/dto/SequenceDetailResponse.java
@@ -22,8 +22,8 @@ public class SequenceDetailResponse {
     @Getter
     @Builder
     public static class SummaryInfo {
-        private Integer totalTrials;
-        private Integer successTrials;
+        private Long totalTrials;
+        private Long successTrials;
         private Double avgAngle;
         private Double inTargetRate;
         private String stabilityLevel;
